+++ example/pages/slider/index.js
@@ -1,0 +1,10 @@
+import Page from '../../common/page';
+
+Page({
+  onChange(event) {
+    wx.showToast({
+      icon: 'none',
+      title: `当前值：${event.detail}`
+    });
+  }
+});