+++ docs/markdown/changelog.md
@@ -1,5 +1,45 @@
 ## 更新日志
 
+### [0.2.1](https://github.com/youzan/vant-weapp/tree/v0.2.1)
+`2018-08-31`
+
+**Improvements**
+
+- 新增 Progress 进度条组件 [\#459](https://github.com/youzan/vant-weapp/pull/459)
+- Field：增加 title-width 属性 [\#482](https://github.com/youzan/vant-weapp/pull/482)
+- Search: 支持在原生 form 组件内使用 [\#480](https://github.com/youzan/vant-weapp/pull/480)
+- Switch: 支持在原生 form 组件内使用 [\#478](https://github.com/youzan/vant-weapp/pull/478)
+- Stepper: 支持在原生 form 组件内使用 [\#477](https://github.com/youzan/vant-weapp/pull/477)
+- SwitchCell: 支持在原生 form 组件内使用 [\#479](https://github.com/youzan/vant-weapp/pull/479)
+- Field: 优化清除按钮颜色 [\#476](https://github.com/youzan/vant-weapp/pull/476)
+- Icon: 新增 class-prefix 属性 [\#475](https://github.com/youzan/vant-weapp/pull/475)
+
+**Bug Fixes**
+
+- 修复 Search focus 属性不生效的问题 [\#484](https://github.com/youzan/vant-weapp/pull/484)
+- 修复 Tabbar z-index 不生效的问题 [\#474](https://github.com/youzan/vant-weapp/pull/474)
+
+
+### [0.2.0](https://github.com/youzan/vant-weapp/tree/v0.2.0)
+`2018-08-22`
+
+**Breaking changes**
+
+- 组件增加 addGlobalClass 属性，支持外部样式覆盖 [\#449](https://github.com/youzan/vant-weapp/pull/449)
+- 优化 Notify 模块导出方式，支持 ES6 模块引入 [\#448](https://github.com/youzan/vant-weapp/pull/448)
+
+**Improvements**
+
+- 新增 Slider 组件 [\#453](https://github.com/youzan/vant-weapp/pull/453)
+- 新增 SwitchCell 组件 [\#455](https://github.com/youzan/vant-weapp/pull/455)
+- 新增 Icon aim 类型 [\#457](https://github.com/youzan/vant-weapp/pull/457)
+- 新增 Field name 属性 [\#450](https://github.com/youzan/vant-weapp/pull/450)
+- 新增 multi-ellipsis 内部样式类 [\#456](https://github.com/youzan/vant-weapp/pull/456)
+
+**Bug Fixes**
+
+- 修复 share 图标无法加粗的问题 [\#457](https://github.com/youzan/vant-weapp/pull/457)
+
 ### [0.1.1](https://github.com/youzan/vant-weapp/tree/v0.1.1)
 `2018-08-21`